refatoração: movendo os componentes  para uma nova pasta e feat: adicionar símbolo de bullet em .text-wrapper
