+++ src/components/FaixasAnimadas/faixaAnimada.css
@@ -20,7 +20,12 @@
     white-space: nowrap;
     color: #fff;
     margin: 0;
-    padding-right: 4px;
+    padding-right: 16px;
+}
+
+.text-wrapper p::after {
+    content: "•";
+    padding-left: 16px;
 }
 
 #text-wrapper1 {
