+++ src/components/FaixasAnimadas/faixaAnimada.jsx
@@ -1,0 +1,34 @@
+import './FaixaAnimada.css';
+
+const FaixaAnimada = () => {
+    return (
+        <section className="faixa">
+            <div className="faixa1">
+                <div className="text-wrapper" id="text-wrapper1">
+                    <p>Construindo o amanhã, hoje</p>
+                    <p>Construindo o amanhã, hoje</p>
+                    <p>Construindo o amanhã, hoje</p>
+                    <p>Construindo o amanhã, hoje</p>
+                    <p>Construindo o amanhã, hoje</p>
+                    <p>Construindo o amanhã, hoje</p>
+                    <p>Construindo o amanhã, hoje</p>
+                    <p>Construindo o amanhã, hoje</p>
+                    <p>Construindo o amanhã, hoje</p>
+                    <p>Construindo o amanhã, hoje</p>
+                    <p>Construindo o amanhã, hoje</p>
+                    <p>Construindo o amanhã, hoje</p>
+                    <p>Construindo o amanhã, hoje</p>
+                    <p>Construindo o amanhã, hoje</p>
+                    <p>Construindo o amanhã, hoje</p>
+                    <p>Construindo o amanhã, hoje</p>
+                    <p>Construindo o amanhã, hoje</p>
+                    <p>Construindo o amanhã, hoje</p>
+                    <p>Construindo o amanhã, hoje</p>
+                    <p>Construindo o amanhã, hoje</p>
+                </div>
+            </div>
+        </section>
+    );
+};
+
+export default FaixaAnimada;
